+++ core/designsystem/src/main/java/com/yapp/breake/core/designsystem/component/Button.kt
@@ -2,14 +2,12 @@ package com.yapp.breake.core.designsystem.component
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
@@ -34,7 +32,6 @@ import com.yapp.breake.core.designsystem.theme.BrakeTheme
 import com.yapp.breake.core.designsystem.theme.Gray200
 import com.yapp.breake.core.designsystem.theme.Gray700
 import com.yapp.breake.core.designsystem.theme.Gray800
-import com.yapp.breake.core.designsystem.theme.KakaoYellow
 import com.yapp.breake.core.designsystem.theme.Red
 import com.yapp.breake.core.designsystem.util.BooleanProvider
 import com.yapp.breake.core.designsystem.util.MultipleEventsCutter
@@ -164,71 +161,6 @@ fun CircleButton(
 	}
 }
 
-@Composable
-fun KakaoLoginButton(
-	modifier: Modifier = Modifier,
-	text: String = "카카오 로그인",
-	onClick: () -> Unit,
-	enabled: Boolean = true,
-) {
-	val multipleEventsCutter = remember { MultipleEventsCutter.get() }
-
-	Button(
-		shape = MaterialTheme.shapes.large,
-		colors = ButtonDefaults.buttonColors(
-			containerColor = KakaoYellow,
-			contentColor = MaterialTheme.colorScheme.onPrimary,
-			disabledContainerColor = Gray700,
-			disabledContentColor = Gray200,
-		),
-		contentPadding = PaddingValues(16.dp),
-		enabled = enabled,
-		onClick = { multipleEventsCutter.processEvent(onClick) },
-		modifier = modifier.fillMaxWidth(),
-	) {
-		Row(
-			modifier = Modifier.fillMaxWidth(),
-			verticalAlignment = Alignment.CenterVertically,
-			horizontalArrangement = Arrangement.Start,
-		) {
-			// 왼쪽 카카오 로고 영역 (전체 폭의 20%)
-			Box(
-				modifier = Modifier
-					.weight(0.2f)
-					.wrapContentHeight(),
-				contentAlignment = Alignment.Center,
-			) {
-				Icon(
-					painter = painterResource(id = R.drawable.ic_kakao_logo),
-					contentDescription = "카카오 로고",
-					tint = MaterialTheme.colorScheme.onPrimary,
-					modifier = Modifier.padding(end = 8.dp),
-				)
-			}
-
-			// 가운데 텍스트 영역 (전체 폭의 60%)
-			Box(
-				modifier = Modifier
-					.weight(0.6f)
-					.wrapContentHeight(),
-				contentAlignment = Alignment.Center,
-			) {
-				Text(
-					text = text,
-					style = BrakeTheme.typography.subtitle16B,
-				)
-			}
-
-			// 오른쪽 빈 공간 영역 (전체 폭의 20%)
-			Box(
-				modifier = Modifier
-					.weight(0.2f)
-					.wrapContentHeight(),
-			)
-		}
-	}
-}
-
 @Preview("Large Buttons")
 @Composable
 private fun LargeButtonPreview(
@@ -275,20 +207,6 @@ private fun CircleButtonPreview() {
 		CircleButton(
 			icon = R.drawable.ic_close,
 			onClick = { },
-		)
-	}
-}
-
-@Preview
-@Composable
-private fun KakaoLoginButtonPreview(
-	@PreviewParameter(BooleanProvider::class) enabled: Boolean,
-) {
-	BrakeTheme {
-		KakaoLoginButton(
-			text = "카카오 로그인",
-			onClick = { },
-			enabled = enabled,
 		)
 	}
 }
