+++ presentation/login/src/main/res/drawable/ic_kakao_logo.xml
